# 🚨 SECURITY: Eliminate pandas.eval() code injection vulnerability (Issue #46)