--- a/src/databeak/models/expression_models.py
+++ b/src/databeak/models/expression_models.py
@@ -1,7 +1,8 @@
-"""Pydantic models for secure mathematical expression validation.
+"""Unified secure expression model for DataBeak mathematical expressions.
 
-This module provides validated models for mathematical expressions used in
-DataBeak column operations, ensuring only safe expressions are accepted.
+This module provides a single, comprehensive SecureExpression type that replaces
+all previous expression types (Formula, Expression, ColumnFormula, ApplyExpression,
+etc.) with a unified model that supports all use cases.
 
 Author: DataBeak Security Team
 Issue: #46 - Address pandas.eval() code injection vulnerability
@@ -17,22 +18,33 @@ from ..utils.secure_evaluator import validate_expression_safety
 
 
 class SecureExpression(BaseModel):
-    """Validated mathematical expression that can be safely evaluated.
+    """Unified secure mathematical expression with validation and context support.
 
-    This model ensures that only safe mathematical operations are allowed
-    in user-provided expressions, preventing code injection attacks.
+    This is the single expression type used throughout DataBeak for all mathematical
+    operations. It provides safety validation, variable substitution, and optional
+    metadata for flexible usage across different scenarios.
 
     Examples of valid expressions:
-        - "col1 + col2"
-        - "abs(col1) * 2.5"
-        - "np.sqrt(col1 + col2)"
-        - "(col1 + col2) / 2"
-        - "max(col1, 100)"
+        - "col1 + col2"                    # Column references
+        - "abs(col1) * 2.5"               # Mathematical functions
+        - "np.sqrt(col1 + col2)"          # Numpy functions
+        - "x * 2 + 10"                    # Variable expressions (for apply operations)
+        - "max(col1, 100)"                # Element-wise operations
 
     Examples of blocked expressions:
-        - "__import__('os').system('rm -rf /')"
-        - "exec('malicious_code')"
-        - "open('/etc/passwd').read()"
+        - "__import__('os').system('rm -rf /')"  # System access
+        - "exec('malicious_code')"               # Code execution
+        - "open('/etc/passwd').read()"           # File access
+
+    Usage patterns:
+        # Simple formula
+        SecureExpression(expression="col1 + col2")
+
+        # Apply operation with variable
+        SecureExpression(expression="x * 2", variable_name="x")
+
+        # Formula with description
+        SecureExpression(expression="col1 + col2", description="Sum of two columns")
     """
 
     model_config = ConfigDict(extra="forbid")
@@ -41,6 +53,16 @@ class SecureExpression(BaseModel):
         description="Safe mathematical expression using column names and mathematical functions",
         min_length=1,
         max_length=1000,  # Prevent extremely long expressions
+    )
+    description: str | None = Field(
+        default=None,
+        description="Optional description of what the expression computes",
+        max_length=200
+    )
+    variable_name: str = Field(
+        default="x",
+        description="Variable name used in expression for apply operations",
+        pattern=r"^[a-zA-Z_][a-zA-Z0-9_]*$"  # Valid Python identifier
     )
 
     @field_validator("expression")
@@ -63,52 +85,6 @@ class SecureExpression(BaseModel):
             raise ValueError(f"Unsafe expression: {e}") from e
         return v
 
-    def __str__(self) -> str:
-        """Return the expression string."""
-        return self.expression
-
-
-class ColumnFormula(BaseModel):
-    """Formula specification for column creation with validation.
-
-    Used in add_column operations where a formula is provided to compute column values from existing
-    columns.
-    """
-
-    model_config = ConfigDict(extra="forbid")
-
-    formula: SecureExpression = Field(
-        description="Mathematical formula referencing existing columns"
-    )
-    description: str | None = Field(
-        default=None,
-        description="Optional description of what the formula computes",
-        max_length=200,
-    )
-
-    def __str__(self) -> str:
-        """Return the formula expression."""
-        return self.formula.expression
-
-
-class ApplyExpression(BaseModel):
-    """Expression specification for apply operations with column context.
-
-    Used in update_column operations where 'x' represents the current column values being
-    transformed.
-    """
-
-    model_config = ConfigDict(extra="forbid")
-
-    expression: SecureExpression = Field(
-        description="Mathematical expression where 'x' represents column values"
-    )
-    variable_name: str = Field(
-        default="x",
-        description="Variable name used in expression to represent column values",
-        pattern=r"^[a-zA-Z_][a-zA-Z0-9_]*$",  # Valid Python identifier
-    )
-
     @field_validator("variable_name")
     @classmethod
     def validate_variable_name(cls, v: str) -> str:
@@ -120,6 +96,8 @@ class ApplyExpression(BaseModel):
     def get_expression_with_column(self, column_name: str) -> str:
         """Get expression with variable substituted for column reference.
 
+        Used for apply operations where a variable represents column values.
+
         Args:
             column_name: Name of the column to substitute for the variable
 
@@ -128,31 +106,74 @@ class ApplyExpression(BaseModel):
         """
         # Safely quote column name for pandas
         safe_column_ref = f"`{column_name.replace('`', '')}`"
-        return self.expression.expression.replace(self.variable_name, safe_column_ref)
+        return self.expression.replace(self.variable_name, safe_column_ref)
 
     def __str__(self) -> str:
         """Return the expression string."""
-        return self.expression.expression
+        return self.expression
+
+    def __repr__(self) -> str:
+        """Return detailed representation."""
+        parts = [f"expression='{self.expression}'"]
+        if self.description:
+            parts.append(f"description='{self.description}'")
+        if self.variable_name != "x":
+            parts.append(f"variable_name='{self.variable_name}'")
+        return f"SecureExpression({', '.join(parts)})"
+
+    # Convenience class methods for common use cases
+    @classmethod
+    def formula(cls, expression: str, description: str | None = None) -> "SecureExpression":
+        """Create expression for column formulas.
+
+        Args:
+            expression: Mathematical expression referencing column names
+            description: Optional description of the formula
+
+        Returns:
+            SecureExpression configured for formula usage
+
+        Example:
+            SecureExpression.formula("col1 + col2", "Sum of columns")
+        """
+        return cls(expression=expression, description=description)
+
+    @classmethod
+    def apply_operation(cls, expression: str, variable: str = "x") -> "SecureExpression":
+        """Create expression for apply operations.
+
+        Args:
+            expression: Mathematical expression using variable (default 'x')
+            variable: Variable name representing column values
+
+        Returns:
+            SecureExpression configured for apply operations
+
+        Example:
+            SecureExpression.apply_operation("x * 2 + 1", "x")
+        """
+        return cls(expression=expression, variable_name=variable)
+
+    @classmethod
+    def condition(cls, expression: str, description: str | None = None) -> "SecureExpression":
+        """Create expression for conditional operations.
+
+        Args:
+            expression: Boolean expression for conditional logic
+            description: Optional description
+
+        Returns:
+            SecureExpression configured for conditional usage
+
+        Example:
+            SecureExpression.condition("col1 > 0", "Positive values")
+        """
+        return cls(
+            expression=expression,
+            description=description or "Conditional expression"
+        )
 
 
-class ConditionalExpression(BaseModel):
-    """Conditional expression for advanced column operations.
-
-    Supports expressions that evaluate to boolean values for filtering or conditional operations.
-    """
-
-    model_config = ConfigDict(extra="forbid")
-
-    condition: SecureExpression = Field(description="Boolean expression for conditional logic")
-    true_value: Any = Field(description="Value to use when condition is True")
-    false_value: Any = Field(description="Value to use when condition is False")
-
-    def __str__(self) -> str:
-        """Return the condition expression."""
-        return self.condition.expression
-
-
-# Type aliases for convenience
+# For backward compatibility - these will be deprecated
 Formula = SecureExpression
 Expression = SecureExpression
-SafeFormula = ColumnFormula