--- a/src/databeak/servers/column_server.py
+++ b/src/databeak/servers/column_server.py
@@ -20,10 +20,15 @@ from ..exceptions import (
     SessionNotFoundError,
 )
 from ..models import OperationType
+from ..models.expression_models import SecureExpression
 from ..models.tool_responses import BaseToolResponse, ColumnOperationResult
+from ..utils.secure_evaluator import SecureExpressionEvaluator
 from .server_utils import get_session_data
 
 logger = logging.getLogger(__name__)
+
+# Global secure expression evaluator instance
+_secure_evaluator = SecureExpressionEvaluator()
 
 # Type aliases
 CellValue = str | int | float | bool | None
@@ -266,8 +271,8 @@ async def add_column(
         Field(description="Single value for all rows or list of values (one per row)"),
     ] = None,
     formula: Annotated[
-        str | None,
-        Field(description="Python expression to compute column values (e.g., 'col1 + col2')"),
+        SecureExpression | str | None,
+        Field(description="Safe mathematical expression to compute column values (e.g., 'col1 + col2')"),
     ] = None,
 ) -> ColumnOperationResult:
     """Add a new column to the dataframe.
@@ -304,8 +309,13 @@ async def add_column(
 
         if formula:
             try:
-                # Use pandas eval to safely evaluate formula
-                df[name] = df.eval(formula)
+                # Convert string to SecureExpression if needed
+                if isinstance(formula, str):
+                    formula = SecureExpression(expression=formula)
+
+                # Use secure evaluator instead of pandas.eval
+                result = _secure_evaluator.evaluate_simple_formula(formula.expression, df)
+                df[name] = result
             except Exception as e:
                 raise InvalidParameterError("formula", formula, f"Invalid formula: {e}") from e
         elif isinstance(value, list):
@@ -619,12 +629,15 @@ async def update_column(
                             "For string operations, use exact expressions: 'x.upper()', 'x.lower()', 'x.strip()', 'x.title()'",
                         )
                 else:
-                    # Use pandas.eval for mathematical expressions
+                    # Use secure evaluator for mathematical expressions
                     try:
-                        # Replace 'x' with the column reference for pandas.eval
-                        safe_expr = expr.replace("x", f"`{column}`")
-                        # pandas.eval is safe - it only allows mathematical operations
-                        df[column] = df.eval(safe_expr, engine="python")
+                        # Create column context mapping for 'x' variable
+                        column_context = {"x": column}
+                        # Use secure evaluator instead of pandas.eval
+                        result = _secure_evaluator.evaluate_column_expression(
+                            expr, df, column_context
+                        )
+                        df[column] = result
                     except Exception as e:
                         raise InvalidParameterError(
                             "expression",
@@ -679,12 +692,15 @@ async def update_column(
                                         "For string operations, use exact expressions: 'x.upper()', 'x.lower()', 'x.strip()', 'x.title()'",
                                     )
                             else:
-                                # Use pandas.eval for mathematical expressions
+                                # Use secure evaluator for mathematical expressions
                                 try:
-                                    # Replace 'x' with the column reference for pandas.eval
-                                    safe_expr = expr.replace("x", f"`{column}`")
-                                    # pandas.eval is safe - it only allows mathematical operations
-                                    df[column] = df.eval(safe_expr, engine="python")
+                                    # Create column context mapping for 'x' variable
+                                    column_context = {"x": column}
+                                    # Use secure evaluator instead of pandas.eval
+                                    result = _secure_evaluator.evaluate_column_expression(
+                                        expr, df, column_context
+                                    )
+                                    df[column] = result
                                 except Exception as e:
                                     raise InvalidParameterError(
                                         "expression",
@@ -761,12 +777,15 @@ async def update_column(
                                         "For string operations, use exact expressions: 'x.upper()', 'x.lower()', 'x.strip()', 'x.title()'",
                                     )
                             else:
-                                # Use pandas.eval for mathematical expressions
+                                # Use secure evaluator for mathematical expressions
                                 try:
-                                    # Replace 'x' with the column reference for pandas.eval
-                                    safe_expr = expr.replace("x", f"`{column}`")
-                                    # pandas.eval is safe - it only allows mathematical operations
-                                    df[column] = df.eval(safe_expr, engine="python")
+                                    # Create column context mapping for 'x' variable
+                                    column_context = {"x": column}
+                                    # Use secure evaluator instead of pandas.eval
+                                    result = _secure_evaluator.evaluate_column_expression(
+                                        expr, df, column_context
+                                    )
+                                    df[column] = result
                                 except Exception as e:
                                     raise InvalidParameterError(
                                         "value",