… pyproject.toml
… src/databeak/services/transformation_operations.py
… tests/security/test_column_server_security.py
… tests/security/test_expression_security.py